buscar funcional

+++ src/main/java/com/erikfora/main/modelos/Resultados.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class Resultados {
 
     @Id
-    long ID;
+    int ID;
 
     @OneToMany()
     @JsonAlias("results")
